Moved components for the cabinet and shop into separate directories
