+++ resources/js/components/Cabinet/Cabinet.vue
