+++ resources/js/components/Shop/Paginator.vue
