+++ resources/js/components/Shop/PurchasesHistory.vue
