+++ resources/js/components/Shop/ThanksModal.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import ModalLayout from './ModalLayout.vue';
+import ModalLayout from '../ModalLayout.vue';
 
 export default {
     name: 'ThanksModal',
